pyproject.toml
[tool.poetry]
name = "codetest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
zeep = "^4.2.1"
pyyaml = "^6.0"
pymysql = "^1.0.3"
cryptography = "^41.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
codetest = "codetest.main.main"
